windows: creating threads under loader lock is a big no-no.
Core initialization should happen inside DllMain, not a secondary thread. The CreateThread() call could fail or deadlock at any time, because it is allowed to take the loader lock.
This is exactly what was happening with the Tasks service async threads.

Instead, all initialization should happen in DllMain, and the Tasks can queue the thread cration on the NWN MainLoop to defer it.